force default feed ids if needed
The user experience when a user doesn't set a feed id for ACARS et. al.
and immediately clicks the 'setup SDR' link is really bad. They don't see
that the feed id is required (they don't get an error), but then once they
assign an SDR the container will fail to start without a useful error
message. While using the site name as part of the feed id isn't ideal,
it at least makes it obvious what happened and is then easy to correct.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>