cleanup: correctly handle variable update in feeder-update
We need to handle the case where this variable doesn't exist (first update from
a version that's a few commits older than this commit), and we need to make
sure that the code doesn't step on itself with the .env file.
So to simplify the logic, have the --update-config run go first (and ensure
that there's a config.json file), and then use jq to update or add that
variable.

Add the missing release notes update.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>