add USB TEMPer style sensor support and rework overall sensor handling
This appears to support quite a few USB temperature sensors. Which should
work on most platforms.

This uses the temper.py file added in the last two commits and rewrites the
temperatuer service quite a bit. It now supports DHT11/22 with either a C
library (seems to work on most systems) or the previously used Python code,
supports BME280 via i2c, and finally adds support for the TEMPer sensor.

To make this more accessible, the UI has been changed to allow the user to
specify which temperature sensor they have and to also declare the GPIO that
they have connected their DHT11/22 to.

In summary, we now support:
- on all systems the TEMPer USB stick
- on Raspberry Pis and the Orange Pi Zero3 the DHT11 and DHT22 (via configurable GPIO)
- on Raspberry Pis the BME280 via i2c (the user needs to manually enable that)

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>