Add Python module that supports TEMPer style USB temperature sensors
The file temper.py comes from https://github.com/ccwienk/temper where it is
provided under the MIT license and can therefore be incorporated here.

In order to use this module in this project, we need to make a few small
modifications. To make it easier to track those they are separated out in the
next commit (so the file added in this commit is the original).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>