change basic setup UI to select protocols up front
This is a rather invasive change into the overall user experience, but
I think it's necessary to move the image a bit further away from its
history of being ADS-B only.

Ask the user to pick the protocols that they want to feed right as
part of the basic setup, and delegate some of the more ADS-B specific
choices until later when they make their data sharing choices.

This commit introduces the concept of the user entering their initials
in order to create good feed names for ACARS et al. Those can of course
be overwritten later on the expert page.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>