move disk-to-run.sh to its own service
run-to-disk stays in adsb-docker so its run after the containers have
been stopped on a regular shutdown
disk-to-run won't do anything if the files in
/run/acars_data already exist
and run-to-disk won't do anything if the files in /run/acars_data are
not there

so this should work fine.

this should fix an issue where a power loss will lead to total acarshub
data loss.
one possible scenario would be that the power loss doesn't stop the
containers like it usually happens on shutdown
on boot, docker service starts and if the /run/acars_data directory
doesn't have the rrd and sqlite files, acarshub is gonna create empty
ones when being started.
adsb-docker starts after docker service, adsb-setup should start at the
same time or hopefully before docker service