fix log-sanitizer sometimes not producing whole log
log-sanitizer was using a shell variable to accumulate all the data
before emitting it
It also used <<< "$VAR" to hand that to a program that sanitized the
log.
My best guess is that this on some systems will limit the log length
somehow.
It's ugly anyhow.
Move to a function that outputs all the various parts of the log on
stdout and pipe that function into the sanitizing programs