add Skystats containers
Just get the YAML files in place and some of the basic variables established,
without much of an attempt of configuring things for now.

We separate out the database as having that on an SD card is not ideal, even
with the 10s writeback frequency... this way you can put it on a different host
and talk to it over the network.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>