Skystats - initial integration
Add a new section to the expert page... we need to figure out what to do about
that page - some of it is 'expert' level stuff, but most is just 'slightly more
advanced' features...

The domestic country is auto populated based on location (well, technically,
based on the country indicator for the closest airport).

The password for the Skystats database is autocreated if not set - but can be
overwritten in the UI if the user wants to spare their SD card and use a remote
PostgreSQL data base.  So all the data base host specific data is there to be
overwritten. But really, maybe that should be hidden by default?

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>