Split Advanced and Expert settings
We were mixing too many different things on the Expert page. Some of them
clearly targeting people who know what they are doing, others far more
accessible to the average user, but simply just the next step after the 'Basic'
settings.

Obviously all of this can be argued about (are temp sensors for experts only?
the bme280 certainly is, but a USB based on is trivial to add) - but this
seemed like a reasonable first stab at which option should go where.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>