add simple recovery app
This should get started both on update and first install, but only for an image.
It runs independently of everything else on port 1089 and allows the user to
downgrade to an earlier version (or even target a version to switch to).

This could of course be abused by someone - but then, if you have access to the
feeder via network, you can already do a lot of really annoying things...

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>