broaden test suite for more of the adsb-setup code
In order to be able to run the tests outside of an installed
app or image, we first need to replace all of the absolute paths
with relative paths so that we can run the code against a test
setup of the file system. That's a fairly invasive change, but
without that testing becomes really challenging.

All paths are now derived from the ADSB_BASE_DIR variable.
utils/paths.py contains all the relative paths.

With that in place, add more unit tests, covering most of the
key modules, as well as some integration tests.

Capture all of this in a GitHub action to automate testing.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>