boot test: add first basic components
this requires a fairly specialized setup to work:

- TFTP / iSCSI server on which this script runs
- existing setup of a Raspberry Pi that can boot from that server
- a Kasa smart plug to power ctcle that RPi

With that in place, you can download a specific ADS-B Feeder image from
GitHub, use that to populate the TFTP / iSCSI server, boot the RPi using
that image (iSCSI allows us to truly simulate having an SD card in
there), and check that the system boots and provides the basis setup UI
(and that it indeed uses the image you expect it to use).

So far this just tests "yes, this comes up and shows the basic setup",
but that would already have caught a couple of the more embarrassing
failures in beta releases.

The next step will be to extend this to test more functionality, given
that the RPi in question of course can have SDR(s) and antenna(s)
connected and we can therefore check that it actually receives data.

Cursor and Claude Sonnet were used to help with some of the
implementation details, but quite a bit of this required several rounds
of rework and a LOT of guidance to actually get to a working state.
Still, the sources show some signs of 'vibe coding'.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>