boot test: keep prepared image and used image separate
During testing the root image will be modified - we should always start
from a clean image. But it makes sense to keep a prepared clean image
around if we make multiple passes on the exact same image (less common
once the test infrastructure is stable, but especially important during
development).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>