github webhook: images should be submitted within minutes
If an image has been around for more than half an hour, it has
definitely been submitted before. No point in passing it on and
confusing the boot test service.

Waiting more than a few minutes makes sense on the off chance of a
system / service reboot, network outage, etc. We want to accept those
release edits for a certain amount of time. But after that it's just
noise.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>