Web UI: implement simple http authentication
This is more like a child lock than any real security. But it's something that
a lot of people have asked for.

You can enable this (and are given a random password to prevent password reuse
issues). Once that is set up, both the regular web UI and roll-back / recovery
functionality require that user name and password.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>